Fixed Bug in Leveling XP
Fixed a bug that allowed for "XP waste" if the XP gained was greater than the XP to the next level would be, after leveling up once. (i.e. you would go from 1 to 2, but then, if your remaining XP was enough to go to 3 or more, it would stall.)